+++ hardware/plugins/PluginProtocols.h
@@ -10,20 +10,20 @@ namespace Plugins {
 		std::vector<byte>	m_sRetainedData;
 		bool m_Secure{ false };
 
-	      public:
+	public:
 		CPluginProtocol() = default;
 		virtual void				ProcessInbound(const ReadEvent* Message);
 		virtual std::vector<byte>	ProcessOutbound(const WriteDirective* WriteMessage);
 		virtual void				Flush(CPlugin* pPlugin, CConnection* pConnection);
-		virtual int					Length() { return m_sRetainedData.size(); };
+		virtual int					Length() { return (int)m_sRetainedData.size(); };
 		virtual bool				Secure() { return m_Secure; };
 
-		static CPluginProtocol *Create(const std::string &sProtocol);
+		static CPluginProtocol* Create(const std::string& sProtocol);
 	};
 
 	class CPluginProtocolLine : CPluginProtocol
 	{
-		void ProcessInbound(const ReadEvent *Message) override;
+		void ProcessInbound(const ReadEvent* Message) override;
 	};
 
 	class CPluginProtocolXML : CPluginProtocol
@@ -31,7 +31,7 @@ namespace Plugins {
 	private:
 		std::string		m_Tag;
 	public:
-	  void ProcessInbound(const ReadEvent *Message) override;
+		void ProcessInbound(const ReadEvent* Message) override;
 	};
 
 	class CPluginProtocolJSON : CPluginProtocol
@@ -39,9 +39,9 @@ namespace Plugins {
 	protected:
 		PyObject* JSONtoPython(Json::Value* pJSON);
 	public:
-	  PyObject *JSONtoPython(const std::string &sJSON);
-	  std::string PythontoJSON(PyObject *pDict);
-	  void ProcessInbound(const ReadEvent *Message) override;
+		PyObject* JSONtoPython(const std::string& sJSON);
+		std::string PythontoJSON(PyObject* pDict);
+		void ProcessInbound(const ReadEvent* Message) override;
 	};
 
 	class CPluginProtocolHTTP : public CPluginProtocol
@@ -49,14 +49,14 @@ namespace Plugins {
 	private:
 		std::string		m_Status;
 		int				m_ContentLength;
-		void*			m_Headers;
+		void* m_Headers;
 		bool			m_Chunked;
 		size_t			m_RemainingChunk;
 	protected:
-		void			ExtractHeaders(std::string*	pData);
-		void Flush(CPlugin *pPlugin, CConnection *pConnection) override;
+		void			ExtractHeaders(std::string* pData);
+		void Flush(CPlugin* pPlugin, CConnection* pConnection) override;
 
-	      public:
+	public:
 		CPluginProtocolHTTP(bool Secure)
 			: m_ContentLength(0)
 			, m_Headers(nullptr)
@@ -65,23 +65,23 @@ namespace Plugins {
 		{
 			m_Secure = Secure;
 		};
-		void ProcessInbound(const ReadEvent *Message) override;
-		std::vector<byte> ProcessOutbound(const WriteDirective *WriteMessage) override;
+		void ProcessInbound(const ReadEvent* Message) override;
+		std::vector<byte> ProcessOutbound(const WriteDirective* WriteMessage) override;
 	};
 
 	class CPluginProtocolWS : public CPluginProtocolHTTP
 	{
 	private:
-		bool	ProcessWholeMessage(std::vector<byte> &vMessage, const ReadEvent * Message);
+		bool	ProcessWholeMessage(std::vector<byte>& vMessage, const ReadEvent* Message);
 	public:
 		CPluginProtocolWS(bool Secure) : CPluginProtocolHTTP(Secure) {};
-		void ProcessInbound(const ReadEvent *Message) override;
-		std::vector<byte> ProcessOutbound(const WriteDirective *WriteMessage) override;
+		void ProcessInbound(const ReadEvent* Message) override;
+		std::vector<byte> ProcessOutbound(const WriteDirective* WriteMessage) override;
 	};
 
 	class CPluginProtocolICMP : CPluginProtocol
 	{
-		void ProcessInbound(const ReadEvent *Message) override;
+		void ProcessInbound(const ReadEvent* Message) override;
 	};
 
 	class CPluginProtocolMQTT : CPluginProtocol
@@ -92,7 +92,7 @@ namespace Plugins {
 	public:
 		CPluginProtocolMQTT(bool Secure) : m_PacketID(1), m_bErrored(false) { m_Secure = Secure; };
 		long MQTTDecodeVariableByte(std::vector<byte>::iterator& pIt);
-		void ProcessInbound(const ReadEvent *Message) override;
-		std::vector<byte> ProcessOutbound(const WriteDirective *WriteMessage) override;
+		void ProcessInbound(const ReadEvent* Message) override;
+		std::vector<byte> ProcessOutbound(const WriteDirective* WriteMessage) override;
 	};
 } // namespace Plugins
